Removed obsolete code, tests cleanup
