+++ src/ImageSharp/Formats/Jpeg/Components/Block8x8F.cs
@@ -768,62 +768,5 @@ namespace SixLabors.ImageSharp.Formats.Jpeg.Components
                 return true;
             }
         }
-
-        /// <summary>
-        /// Returns index of the last non-zero element in this matrix.
-        /// </summary>
-        /// <returns>
-        /// Index of the last non-zero element. Returns -1 if all elements are equal to zero.
-        /// </returns>
-        [MethodImpl(InliningOptions.ShortMethod)]
-        public int GetLastNonZeroIndex()
-        {
-#if SUPPORTS_RUNTIME_INTRINSICS
-            if (Avx2.IsSupported)
-            {
-                const int equalityMask = unchecked((int)0b1111_1111_1111_1111_1111_1111_1111_1111);
-
-                Vector256<int> zero8 = Vector256<int>.Zero;
-
-                ref Vector256<float> mcuStride = ref Unsafe.As<Block8x8F, Vector256<float>>(ref this);
-
-                for (int i = 7; i >= 0; i--)
-                {
-                    int areEqual = Avx2.MoveMask(Avx2.CompareEqual(Avx.ConvertToVector256Int32WithTruncation(Unsafe.Add(ref mcuStride, i)), zero8).AsByte());
-
-                    if (areEqual != equalityMask)
-                    {
-                        // Each 4 bits represents comparison operation for each 4-byte element in input vectors
-                        // LSB represents first element in the stride
-                        // MSB represents last element in the stride
-                        // lzcnt operation would calculate number of zero numbers at the end
-
-                        // Given mask is not actually suitable for lzcnt as 1's represent zero elements and 0's represent non-zero elements
-                        // So we need to invert it
-                        int lzcnt = BitOperations.LeadingZeroCount(~(uint)areEqual);
-
-                        // As input number is represented by 4 bits in the mask, we need to divide lzcnt result by 4
-                        // to get the exact number of zero elements in the stride
-                        int strideRelativeIndex = 7 - (lzcnt / 4);
-                        return (i * 8) + strideRelativeIndex;
-                    }
-                }
-
-                return -1;
-            }
-            else
-#endif
-            {
-                int index = Size - 1;
-                ref float elemRef = ref Unsafe.As<Block8x8F, float>(ref this);
-
-                while (index >= 0 && (int)Unsafe.Add(ref elemRef, index) == 0)
-                {
-                    index--;
-                }
-
-                return index;
-            }
-        }
     }
 }
